refactor: Rename ConsistencyWorker to HTTPWorker
